+++ src/allow.rs
@@ -5,7 +5,7 @@ use std::iter::Peekable;
 
 use crate::errors::CompilationError;
 use crate::lex::{Token, TokenType};
-use crate::ptypes::{AllowClause, AttrDomain, Attribute, Class, ClassFlavor, Clause, Signal};
+use crate::ptypes::{AllowClause, AttrDomain, Attribute, Class, ClassFlavor, Clause, FPos, Signal};
 use crate::putil;
 use crate::zpl;
 
@@ -22,7 +22,7 @@ struct ParseAllowState {
 impl ParseAllowState {
     fn new(root_tok: Token) -> ParseAllowState {
         ParseAllowState {
-            root_tok,
+            root_tok: root_tok.clone(),
             ..Default::default()
         }
     }
@@ -41,7 +41,7 @@ impl ParseAllowState {
     ///
     /// The `server` vector in the [AllowClause] will just have one element -- a service clause
     /// that may have attributes from other domains (eg, endpoint or user attributes).
-    fn to_allow_clause(&mut self, clause_id: usize) -> AllowClause {
+    fn to_allow_clause(&mut self, clause_id: usize, last_tok: Token) -> AllowClause {
         let mut client_user_clause = self.client_user_clause.take().unwrap_or(Clause::new(
             ClassFlavor::User,
             zpl::DEF_CLASS_USER_NAME,
@@ -111,6 +111,10 @@ impl ParseAllowState {
         }
         AllowClause {
             clause_id,
+            span: (
+                self.root_tok.clone().into(),
+                FPos::new(last_tok.line, last_tok.col + last_tok.size - 1),
+            ),
             client: client_clauses,
             // TODO: Note that above we spend a lot of time to move the attributes on each
             // client class into the client class of the attribute domain. We do not to that (yet?)
@@ -155,7 +159,7 @@ pub fn parse_allow(
     let mut ps = PState::new(&parse_state.root_tok);
 
     // To parse this we start parsing and break if we hit a ON or a TO.
-    ps.parse_tags_attrs_and_classname(
+    let _ = ps.parse_tags_attrs_and_classname(
         &mut tokens,
         classes_idx,
         &ParseOpts::stop_at_any(&[TokenType::To, TokenType::On]),
@@ -268,7 +272,7 @@ pub fn parse_allow(
                 ));
             }
             // pop the TO off, leaving the 'access'.
-            putil::require_tt(
+            let _ = putil::require_tt(
                 &parse_state.root_tok,
                 tokens.next(),
                 "TO",
@@ -289,13 +293,15 @@ pub fn parse_allow(
 
     // The remaining tokens should start with "access ..." or "signal ..." which we pass to the service class parser.
     // If there is a signal token at the end, only that token will remain after this function.
-    parse_allow_service_clause(&mut parse_state, &mut tokens, classes_idx, classes_map)?;
+    let mut last_tok =
+        parse_allow_service_clause(&mut parse_state, &mut tokens, classes_idx, classes_map)?;
 
     if let Some(_tok) = tokens.peek() {
-        parse_allow_signal_clause(&mut parse_state, &mut tokens, classes_idx, classes_map)?;
+        last_tok =
+            parse_allow_signal_clause(&mut parse_state, &mut tokens, classes_idx, classes_map)?;
     }
 
-    let mut ac = parse_state.to_allow_clause(statement_id);
+    let mut ac = parse_state.to_allow_clause(statement_id, last_tok);
 
     // Set any UNSPECIFIED (lacking domain) attributes to the domain of the clause they are in.
 
@@ -412,17 +418,19 @@ where
 /// There may also be a signal clause after the service clause. If there is a
 /// signal clause, tokens will remain in the queue after this function, otherwise
 /// this function will process all the tokens.
+///
+/// On a successful parse this returns the last token parsed.
 fn parse_allow_service_clause<'a, I>(
     pa_state: &mut ParseAllowState,
     tokens: &mut Peekable<I>,
     classes_idx: &HashMap<String, String>,
     classes_map: &HashMap<String, Class>,
-) -> Result<(), CompilationError>
+) -> Result<Token, CompilationError>
 where
     I: Iterator<Item = &'a Token>,
 {
     // Pop off the "ACCESS" token...
-    putil::require_tt(
+    let _ = putil::require_tt(
         &pa_state.root_tok,
         tokens.next(),
         "ACCESS",
@@ -434,7 +442,8 @@ where
     let mut ps = PState::new(&pa_state.root_tok);
     // Signal clause will always
     let popts = ParseOpts::stop_at_any(&[TokenType::On, TokenType::Eos, TokenType::Signal]);
-    ps.parse_tags_attrs_and_classname(tokens, classes_idx, &popts, "service clause")?;
+    let mut last_token =
+        ps.parse_tags_attrs_and_classname(tokens, classes_idx, &popts, "service clause")?;
 
     let cn = ps.class_name.as_ref().unwrap();
     if classes_map.get(cn).unwrap().flavor != ClassFlavor::Service {
@@ -459,7 +468,7 @@ where
 
                 let mut nested_ps = PState::new(&pa_state.root_tok);
 
-                nested_ps.parse_tags_attrs_and_classname(
+                last_token = nested_ps.parse_tags_attrs_and_classname(
                     tokens,
                     classes_idx,
                     &ParseOpts::stop_at_any(&[TokenType::Eos, TokenType::Signal]),
@@ -517,23 +526,26 @@ where
     }
 
     pa_state.service_clause = Some(service_clause);
-    Ok(())
+    Ok(last_token)
 }
 
 // Expects a signal clause of the form SIGNAL <STRING> TO <SERVICE>
 // Since a signal clause is the final clause of the allow statement,
 // we expect the EOS token, thus the queue will be empty after execution.
+//
+// On successful parse this returns the last token parsed and the tokens
+// iterator is empty.
 fn parse_allow_signal_clause<'a, I>(
     pa_state: &mut ParseAllowState,
     tokens: &mut Peekable<I>,
     _classes_idx: &HashMap<String, String>,
     classes_map: &HashMap<String, Class>,
-) -> Result<(), CompilationError>
+) -> Result<Token, CompilationError>
 where
     I: Iterator<Item = &'a Token>,
 {
     // Pop off the SIGNAL token
-    putil::require_tt(
+    let _ = putil::require_tt(
         &pa_state.root_tok,
         tokens.next(),
         "SIGNAL",
@@ -565,7 +577,7 @@ where
     }
 
     // The next part should be the token TO
-    putil::require_tt(
+    let _ = putil::require_tt(
         &pa_state.root_tok,
         tokens.next(),
         "TO",
@@ -573,8 +585,11 @@ where
         TokenType::To,
     )?;
 
-    // The final portion of the signal clause should be an existing service class
+    let last_tok: Token;
+
+    // The final portion of the signal clause must be an existing service class
     if let Some(tok) = tokens.next() {
+        last_tok = tok.clone();
         // The service does not share a name with a reserved keyword
         if let TokenType::Literal(ref service_name) = tok.tt {
             // We require the requested service to exist in the list of services
@@ -623,7 +638,7 @@ where
     }
 
     pa_state.signal_clause = Some(Signal::new(message, target));
-    return Ok(());
+    return Ok(last_tok);
 }
 
 // Note that this does not check for duplicates.
@@ -721,16 +736,19 @@ impl PState {
     }
 
     /// Parse a class and its attributes.
+    ///
+    /// On a successful parse, this returns the last token parsed.
     fn parse_tags_attrs_and_classname<'a, I>(
         &mut self,
         tokens: &mut Peekable<I>,
         classes: &HashMap<String, String>,
         opts: &ParseOpts,
         context: &str,
-    ) -> Result<(), CompilationError>
+    ) -> Result<Token, CompilationError>
     where
         I: Iterator<Item = &'a Token>,
     {
+        let mut last_token = Token::default();
         let mut tcount = 0;
         let mut break_count = 0;
         while let Some(tokref) = tokens.peek() {
@@ -744,13 +762,13 @@ impl PState {
             match &tokref.tt {
                 TokenType::And | TokenType::Comma => {
                     // These are delimiter tokens.
-                    tokens.next();
+                    last_token = tokens.next().unwrap().clone();
                 }
                 TokenType::Tuple((name, value)) => {
                     // This is an attribute.
                     let attr = Attribute::attr_domain_opt(name, value);
                     self.attrs.push(attr);
-                    tokens.next();
+                    last_token = tokens.next().unwrap().clone();
                 }
                 TokenType::Literal(s) => {
                     // This could be a class name or a tag name.
@@ -767,9 +785,10 @@ impl PState {
                         self.class_name = Some(class.clone());
                         let tok = tokens.next().unwrap();
                         self.class_name_token = Some(tok.clone());
+                        last_token = tok.clone();
                     } else {
                         self.attrs.push(Attribute::tag_domain_opt(s));
-                        tokens.next();
+                        last_token = tokens.next().unwrap().clone();
                     }
                 }
                 TokenType::With => {
@@ -811,7 +830,7 @@ impl PState {
             ));
         }
 
-        Ok(())
+        Ok(last_token)
     }
 }